source/bend/app/requirements.txt
django
psycopg2-binary
requests

# For OAuth sign-ins
google-auth
django-allauth
dj-rest-auth
djangorestframework
social-auth-app-django
django-cors-headers
djangorestframework-simplejwt